+++ ext/agoo/request.c
@@ -554,7 +554,7 @@ request_init(VALUE mod) {
     rack_run_once_val = rb_str_new_cstr("rack.run_once");	rb_gc_register_address(&rack_run_once_val);
     rack_url_scheme_val = rb_str_new_cstr("rack.url_scheme");	rb_gc_register_address(&rack_url_scheme_val);
     rack_version_val = rb_str_new_cstr("rack.version");		rb_gc_register_address(&rack_version_val);
-    rack_version_val_val = rb_str_new_cstr("2.0.3");		rb_gc_register_address(&rack_version_val_val);
+    rack_version_val_val = rb_str_new_cstr("1.3");		rb_gc_register_address(&rack_version_val_val);
     request_method_val = rb_str_new_cstr("REQUEST_METHOD");	rb_gc_register_address(&request_method_val);
     script_name_val = rb_str_new_cstr("SCRIPT_NAME");		rb_gc_register_address(&script_name_val);
     server_name_val = rb_str_new_cstr("SERVER_NAME");		rb_gc_register_address(&server_name_val);