+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### 1.0.0 - 2018-02-06
+
+	- Add not found handler support.
+	- Fixed segfault due to GC of handlers.
+	
 ### 0.9.1 - 2018-01-28
 
 Updated to get travis working correctly.